Shape Analysis pass
This patch adds a shape analysis pass that dumps TPP shapes
and surrounding affineparallel op shapes in a file which
can be chosen by setting the environment as follows:
PLAIDML_SHAPE_ANALYSIS_OUTPUT=<path>. The result file
contains annotations for the surrounding loop nest of
a tpp operation marked as PARALLEL if the iteration level
is marked to run in parallel, SERIAL otherwise. This annotation
is followed by (<lowerbound>), (<upperbound>), (<steps>) information.